Upstream Update
8e310303903792c010d03c046ba75f8b18f7d3a7
Add support for OSC color sequences

273db5ceaf392e68c2faf8f7dec14ea2e25e980d
follow-up fix for OSC color sequences, return

Otherwise the message "erresc: unknown str" is printed.

a0467c802d4f86ed162486e3453dd61181423902
Fix null pointer access in strhandle

According to the spec the argument is optional for 104, so p can be
NULL as can be tested with printf '\x1b]104\x07'. This is a regression
of 8e31030.